# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "教室排课、教室资源管理和冲突检测 RESTful API。",
        "title": "教室排课助手 API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/schedules": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "List timetable entries",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "week",
                        "name": "week",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "class id",
                        "name": "class_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "teacher id",
                        "name": "teacher_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "classroom id",
                        "name": "classroom_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/adjustments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "List adjustment history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/schedules/export": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Export a timetable as JSON or CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "class|teacher|classroom",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "entity id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "week",
                        "name": "week",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "json|csv",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/generate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Generate a timetable",
                "parameters": [
                    {
                        "description": "scheduling input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.GenerateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/move": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Move a timetable entry to a free slot",
                "parameters": [
                    {
                        "description": "move payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.MoveScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/swap": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Swap two timetable entries",
                "parameters": [
                    {
                        "description": "swap payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.SwapScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Get one timetable entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "schedule id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gbschedule_gbschedule_internal_dto.CourseRequirement": {
            "type": "object",
            "required": [
                "course_id",
                "weekly_periods"
            ],
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "consecutive": {
                    "type": "boolean"
                },
                "course_id": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "integer"
                },
                "weekly_periods": {
                    "type": "integer",
                    "maximum": 40,
                    "minimum": 1
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.GenerateScheduleRequest": {
            "type": "object",
            "required": [
                "courses",
                "days_per_week",
                "periods_per_day",
                "weeks"
            ],
            "properties": {
                "class_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "classroom_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "courses": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.CourseRequirement"
                    }
                },
                "days_per_week": {
                    "type": "integer",
                    "maximum": 7,
                    "minimum": 1
                },
                "periods_per_day": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1
                },
                "semester": {
                    "type": "string",
                    "maxLength": 128
                },
                "teacher_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "weeks": {
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.MoveScheduleRequest": {
            "type": "object",
            "required": [
                "classroom_id",
                "day_of_week",
                "schedule_id",
                "time_slot_id",
                "week"
            ],
            "properties": {
                "classroom_id": {
                    "type": "integer"
                },
                "day_of_week": {
                    "type": "integer",
                    "maximum": 7,
                    "minimum": 1
                },
                "schedule_id": {
                    "type": "integer"
                },
                "time_slot_id": {
                    "type": "integer"
                },
                "week": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_gbschedule_gbschedule_internal_dto.SwapScheduleRequest": {
            "type": "object",
            "required": [
                "schedule_a_id",
                "schedule_b_id"
            ],
            "properties": {
                "schedule_a_id": {
                    "type": "integer"
                },
                "schedule_b_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}